src/main/java/uk/ac/glasgow/jagora/BuyOrder.java
package uk.ac.glasgow.jagora;

import uk.ac.glasgow.jagora.trader.Trader;
import uk.ac.glasgow.jagora.world.TickEvent;

public abstract class BuyOrder extends Order implements Comparable<BuyOrder> {

	public BuyOrder(Trader trader, Stock stock, Integer quantity) {
		super(trader, stock, quantity);
	}

	@Override
	public void satisfyTrade(TickEvent<Trade> executedTrade) throws TradeExecutionException {
		trader.buyStock(executedTrade.event);		
		tradeHistory.add(executedTrade);
	}

	@Override
	public void rollBackTrade(TickEvent<Trade> executedTrade) throws TradeExecutionException {
		if (tradeHistory.remove(executedTrade))
			trader.sellStock(executedTrade.event);
	}
	
	public abstract Double getPrice();
	
	@Override
	public int compareTo(BuyOrder order) {
		return order.getPrice().compareTo(this.getPrice());
	}

}